Develop

See merge request wangxiao_xueyanandroid_business/livevideo!412

+++ src/main/java/com/xueersi/parentsmeeting/modules/livevideo/studyreport/business/StudyReportAction.java
@@ -13,7 +13,7 @@ public interface StudyReportAction extends LiveProvide {
      *
      * @param uid
      */
-    void onFirstRemoteVideoDecoded(int uid);
+    void onFirstRemoteVideoDecoded(long uid);
 
     /**
      * 接麦用户进入
@@ -21,7 +21,7 @@ public interface StudyReportAction extends LiveProvide {
      * @param uid
      * @param elapsed
      */
-    void onUserJoined(int uid, int elapsed);
+    void onUserJoined(long uid, int elapsed);
 
     /**
      * 接麦用户退出
@@ -29,7 +29,7 @@ public interface StudyReportAction extends LiveProvide {
      * @param uid
      * @param reason
      */
-    void onUserOffline(int uid, int reason);
+    void onUserOffline(long uid, int reason);
 
     /**
      * 截图
